src/requirements.txt
selenium==3.141.0
pytest==6.2.0
requests==2.26.0
allure-pytest==2.9.42
pytest-xdist==2.3.0
